+++ Server/CoursProject/src/main/java/com/primavera/CoursProject/application/daos/TransactionDAO.java
@@ -1,6 +1,6 @@
 package com.primavera.CoursProject.application.daos;
 
 public interface TransactionDAO {
-	public String buyBitcoins(String brokerId, int bitcoins, int money);
+	public String buyBitcoins(String brokerId, double bitcoins, double money);
 	public void addTransaction(String purchaseId, String brokerId);
 }
